Clarify application is for Windows only
As confusion rose in https://github.com/grocy/grocy-desktop/issues/49. I may have overread that, but it is also somewhat ambiguous.

> A (Windows) desktop application

means:

> A Windows desktop application

and/or

> A desktop application

This could e.g. mean the application is not yet ready for macOS or Linux and may need work. (And Windows is just the current compatibility.)
Or, as in this case, it is Windows only. I just would not put Windows in brackets then, as it actually is a "Windows desktop application" then.

We could even replace it with the specific framework or so if we want to be more clear.